feat(architecture): restrict circuit_breaker module visibility
- Convert `pub mod circuit_breaker` to `pub(crate) mod circuit_breaker` in `src/run/swebench.rs` to prevent internal abstraction leaks.
- Relocate unit tests for the circuit breaker from `tests/swebench_systemic_halt.rs` (which tests the public API) to a `#[cfg(test)]` block in `src/run/swebench.rs` since integration tests cannot access crate-private modules.
